+++ d3textdb/src/d3textdb/__init__.py
@@ -12,7 +12,7 @@ except ModuleNotFoundError:
     pass
 
 from .d3textdb import D3TextDB, OntologyInUseError
-from .owl import ParsedOntology, ParsedProperty, ParsedTriple, parse_owl
+from .owl import OntologyMetadata, ParsedOntology, ParsedProperty, ParsedTriple, parse_owl, peek_ontology_metadata
 
 try:
     import stackprinter
@@ -31,8 +31,10 @@ except ModuleNotFoundError:
 
 __all__ = [
     "D3TextDB",
+    "OntologyMetadata",
     "ParsedOntology",
     "ParsedProperty",
     "ParsedTriple",
     "parse_owl",
+    "peek_ontology_metadata",
 ]